fixes wirecutters

+++ code/game/objects/items/tools/wirecutters.dm
@@ -30,6 +30,8 @@
 		icon_state = "cutters_[param_color]"
 
 /obj/item/wirecutters/attack(mob/living/carbon/C, mob/user)
+	if(tearoutteeth(C, user)) //HIPPIE CHANGE
+		return FALSE //HIPPIE CHANGE
 	if(istype(C) && C.handcuffed && istype(C.handcuffed, /obj/item/restraints/handcuffs/cable))
 		user.visible_message("<span class='notice'>[user] cuts [C]'s restraints with [src]!</span>")
 		qdel(C.handcuffed)
